app/requirements.txt
# --- Core Web Framework ---
fastapi==0.110.0
uvicorn[standard]==0.23.2

# --- Pydantic / Environment Management ---
pydantic==1.10.14
python-dotenv==1.0.1
# (pydantic-settings requires Python>=3.10; remove it and use plain python-dotenv)

# --- HTTP Clients ---
httpx==0.27.0
requests==2.31.0
aiofiles==23.2.1

# --- Machine Learning / NLP ---
transformers==4.35.2         # works with PyTorch 2.1 on Python 3.9
torch==2.1.0
sentencepiece==0.1.99
protobuf==4.25.3
numpy==1.26.4
tqdm==4.66.2

# --- Logging / Developer Tools ---
loguru==0.7.2
rich==13.7.1

# --- CORS / Multipart / Security ---
python-multipart==0.0.6
passlib[bcrypt]==1.7.4

# --- Optional (Testing & Utilities) ---
pytest==7.4.3
